move #142: post, post-id-message 폴더구조 수정
- 프로필 이미지 기본값 변경
- 각각 페이지에서 사용하는 constants 이동 분리
