+++ src/components/post-id-message-page/content-textarea/ContentTextarea.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .toastui-editor-defaultUI .ProseMirror {
   padding: 18px 25px;
