+++ src/components/post-id-message-page/custom-drop-down/CustomDropdown.scss
@@ -1,4 +1,4 @@
-@import '../../styles//variables';
+@import '../../../styles//variables';
 
 .message-form__arrow-icon {
   width: 16px;
