+++ src/components/post-id-message-page/profile-options/ProfileOptions.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .message-form__title {
   @include font-style(24, Bold);
