+++ src/components/post-id-message-page/sender-input/SenderInput.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .message-form__sender {
   input {
