+++ src/components/post-page/color-picker/ColorPicker.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 @mixin flex-center {
   display: flex;
