+++ src/components/post-page/image-picker/ImagePicker.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables', '../../styles/reset';
+@import '../../../styles/variables';
 
 @mixin flex-center {
   display: flex;
